refactor: move error logging to bug tracking systems into separate custom winston Transport
