+++ package.json
@@ -88,6 +88,7 @@
     "qs": "^6.8.0",
     "rollbar": "^2.7.1",
     "throng": "^4.0.0",
+    "triple-beam": "^1.3.0",
     "twilio": "^3.35.0",
     "uuid": "^8.1.0",
     "web3": "^2.0.0-alpha.1",
@@ -95,7 +96,8 @@
     "web3-providers": "^2.0.0-alpha.1",
     "web3-utils": "^2.0.0-alpha.1",
     "websocket": "^1.0.29",
-    "winston": "^3.2.1"
+    "winston": "^3.2.1",
+    "winston-transport": "^4.4.0"
   },
   "devDependencies": {
     "@babel/core": "^7.4.5",
