+++ src/server/server-middlewares.js
@@ -12,7 +12,8 @@ import addVerificationMiddlewares from './verification/verificationAPI'
 import addSendMiddlewares from './send/sendAPI'
 import addLoadTestMiddlewares from './loadtest/loadtest-middleware'
 import addCypressMiddleware from './cypress/cypress-middleware'
-import { rollbar, addRequestLogger } from '../imports/logger'
+import { addRequestLogger } from '../imports/logger'
+import { rollbar } from '../imports/logger/ErrorsTransport'
 import VerificationAPI from './verification/verification'
 import createDisposeEnrollmentsTask from './verification/cron/DisposeEnrollmentsTask'
 import addClaimQueueMiddlewares from './claimQueue/claimQueueAPI'
